requirements.txt
click>=8.1.0
loguru>=0.7.0
pyyaml>=6.0
pydantic>=2.0.0
Pillow>=10.0.0
